+++ auth.ts
@@ -1,0 +1,51 @@
+import NextAuth from 'next-auth';
+import Credentials from 'next-auth/providers/credentials';
+import { authConfig } from './auth.config';
+import { prisma } from '@/lib/prisma';
+import bcrypt from 'bcryptjs';
+import { z } from 'zod';
+
+export const { auth, signIn, signOut, handlers } = NextAuth({
+  ...authConfig,
+  secret: process.env.AUTH_SECRET,
+  providers: [
+    Credentials({
+      async authorize(credentials) {
+        const parsedCredentials = z
+          .object({ username: z.string(), password: z.string().min(4) })
+          .safeParse(credentials);
+
+        if (parsedCredentials.success) {
+          const { username, password } = parsedCredentials.data;
+          
+          const user = await prisma.user.findUnique({
+            where: { username },
+          });
+
+          if (!user) return null;
+
+          const passwordsMatch = await bcrypt.compare(password, user.password);
+
+          if (passwordsMatch) return { id: user.id, name: user.name, username: user.username };
+        }
+
+        console.log('Credenciais inválidas');
+        return null;
+      },
+    }),
+  ],
+  callbacks: {
+    jwt({ token, user }) {
+      if (user) {
+        token.username = (user as any).username;
+      }
+      return token;
+    },
+    session({ session, token }) {
+      if (session.user) {
+        (session.user as any).username = token.username;
+      }
+      return session;
+    },
+  },
+});